requirements.txt
streamlit

transformers

torch
torchvision

paddlepaddle

paddleocr

pymupdf

pandas

openpyxl

opencv-python-headless
Pillow

sentencepiece
accelerate
einops

transformers[torch]

bitsandbytes